Added testBitfieldValuesValidation() method to TestBitifleds.java class. This method will:
a.  Check read/write operation on suppported bitfields by writing a valid value and then reading it from barinboard to verify value was indeed written

b. Write invalid values and then reading it from brainboard to verify that the default value is returned isntead (because we can't write invalid values)

On this commit, only item "a" has been implemented

Note: It will be good to have an exception handle item "b" on the SparkyAndroidLib

Also modified HandleCmd methods to return "doubles" instead of "ints" for ease of handling data and comparison